requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asynco==0.0.5
asyncpg==0.22.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
DateTime==4.3
discord==1.0.1
discord.py==1.5.1
docopt==0.6.2
idna==2.10
multidict==4.7.6
PIL-Tools==1.0
Pillow==8.1.2
pipreqs==0.4.10
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
requests==2.25.1
samp-client==3.0.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarg==0.1.9
yarl==1.5.1
youtube-dl==2021.1.8
zope.interface==5.2.0
